+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `NavisworksUtils` to support `NavisworksApplication` and `NavisworksPlatform`.
 - Add `AppNameSpace` and `UpgradeCode` to `PackageBuilder`.
 - Update `ProductCode` and `UpgradeCode` to convert `Guid` to `ToStringBraces`.
+- Update `AppVersion` to convert `Version` to `ToString(3)`.
 ### Tests
 - Add `DataBuild_Tests` to test `DataBuilder`
 - Add `PackageContentsBuilder_AutoCAD_Tests` and `PackageContentsBuilder_Revit_Tests`.